 requirements.txt
speechrecognition
pyttsx3
requests
python-dotenv
